+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To be defined.
 
+## [0.0.9] - 2022-02-25
+### Changed
+- Standardizes WithXYZOptions to set params instead of adding.
+- `New` is basic server, `NewDefault` is loaded one.
+  
 ## [0.0.8] - 2022-02-25
 ### Changed
 - Removes any external dependency on `expvar`.